refactor(exceptions): make exceptions extensible
Next step should be to utilize exceptions instead of `assert`. The main
advantage of it is that you can catch specific exception instead of
`AssertionError`.
