+++ generals/__init__.py
@@ -1,9 +1,11 @@
 from gymnasium.envs.registration import register
 
 from generals.agents import AgentFactory
+from generals.core.exceptions import GeneralsBotError
 from generals.core.grid import Grid, GridFactory
 from generals.core.replay import Replay
 from generals.envs.pettingzoo_generals import PettingZooGenerals
+from generals.remote.exceptions import GeneralsIOClientError, RegisterAgentError
 
 __all__ = [
     "AgentFactory",
@@ -11,6 +13,9 @@ __all__ = [
     "PettingZooGenerals",
     "Grid",
     "Replay",
+    "GeneralsBotError",
+    "GeneralsIOClientError",
+    "RegisterAgentError",
 ]
