+++ generals/core/exceptions.py
@@ -1,0 +1,5 @@
+
+class GeneralsBotError(Exception):
+    """Base generals-bot exception """
+
+    pass
