+++ generals/remote/generalsio_client.py
@@ -23,30 +23,7 @@ def autopilot(agent_id: str, user_id: str, lobby_id: str) -> None:
             if client.status == "lobby":
                 client.join_game()
 
-
-class GeneralsBotError(Exception):
-    """Base generals-bot exception
-    TODO: find a place for exceptions
-    """
-
-    pass
-
-
-class GeneralsIOClientError(GeneralsBotError):
-    """Base GeneralsIOClient exception"""
-
-    pass
-
-
-class RegisterAgentError(GeneralsIOClientError):
-    """Registering bot error"""
-
-    def __init__(self, msg: str) -> None:
-        super().__init__()
-        self.msg = msg
-
-    def __str__(self) -> str:
-        return f"Failed to register the agent. Error: {self.msg}"
+from .exceptions import GeneralsIOClientError, RegisterAgentError
 
 
 class GeneralsIOClient(SimpleClient):
